+++ live-ingester/src/main/java/org/alfresco/hxi_connector/live_ingester/adapters/storage/connector/PreSignedUrlRequester.java
@@ -41,8 +41,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import org.alfresco.hxi_connector.live_ingester.adapters.storage.StorageLocationRequest;
-import org.alfresco.hxi_connector.live_ingester.adapters.storage.StorageLocationRequester;
 import org.alfresco.hxi_connector.live_ingester.domain.exception.LiveIngesterRuntimeException;
 
 @Component
