+++ live-ingester/src/main/java/org/alfresco/hxi_connector/live_ingester/adapters/storage/connector/StorageLocationRequest.java
@@ -23,10 +23,7 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.hxi_connector.live_ingester.adapters.storage;
+package org.alfresco.hxi_connector.live_ingester.adapters.storage.connector;
 
-import java.io.InputStream;
-import java.net.URL;
-
-public record FileUploadRequest(InputStream inputStream, String contentType, URL storageLocation)
+public record StorageLocationRequest(String nodeId, String contentType)
 {}
