+++ live-ingester/src/main/java/org/alfresco/hxi_connector/live_ingester/adapters/storage/connector/StorageLocationRequester.java
@@ -23,7 +23,11 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.hxi_connector.live_ingester.adapters.storage;
+package org.alfresco.hxi_connector.live_ingester.adapters.storage.connector;
 
-public record StorageLocationRequest(String nodeId, String contentType)
-{}
+import java.net.URL;
+
+public interface StorageLocationRequester
+{
+    URL requestStorageLocation(StorageLocationRequest storageLocationRequest);
+}
