+++ live-ingester/src/test/java/org/alfresco/hxi_connector/live_ingester/adapters/storage/connector/HttpFileUploaderTest.java
@@ -47,7 +47,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import org.alfresco.hxi_connector.live_ingester.adapters.storage.FileUploadRequest;
 import org.alfresco.hxi_connector.live_ingester.domain.exception.LiveIngesterRuntimeException;
 
 class HttpFileUploaderTest
